refactor(v2): extract shared typed repository runtime
